backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "pyjwt>=2.10.1",
    "click>=8.1.0",
    "werkzeug>=3.0.0",
]
